+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Pico-Writer
 
 v1.0.1 ()
 - Added view mode
+- Improved changes detection before quiting
+- Added install uninstall scripts
 
 v1.0.0 (28-10-2025)
 - Launch